Some work on fleshing out biome and dimension allow/deny lists
Final resolution of spawns - which need occur, necessarily, at the end
of the block registry event (the start of the item registry event, at a
minimum) is the next step. Will need to create a BlockMatcher setup to
take over for the Replacements and a generic utility for turning
OS4BlockData into a BlockState.